New Deleted directory.  Reviewed mIRC pages.
The new /deleted/ directory is where deleted pages go to die or live in
everlasting pergatory.  They should contain no archival value nor
historic interest.  The subdirectory /materials for reuse/ is where
pages which contain materials that should be reused should visit before
they end up in permanent deletion pergatory.

This system is a part of the upgrade from the old static site with pages
and content dating back to 1993.