Renamed /deleted/ to /purgatory/
Files that no longer fit on the website and aren't being specifically
maintained for archival and historic record, but offer some contents
worthy of retention and encorporation into other articles.  Feel free to
contribute to the site by recycling the materials within, and provide
credit where credit is due to authors of IRCHelp's past.